+++ jest.setup.ts
@@ -1,0 +1,5 @@
+import { ComponentRegister } from "./src/ComponentRegister";
+
+afterEach(() => {
+  ComponentRegister.reset();
+});